+++ packages/arcade_cache/lib/arcade_cache.dart
@@ -1,1 +1,2 @@
 export 'src/cache_manager.dart' show BaseCacheManager;
+export 'src/memory_cache_manager.dart' show MemoryCacheManager;