+++ packages/arcade_cache/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.0.3
+
+ - **FEAT**: add MemoryCacheManager implementation for in-memory caching with TTL support.
+
 ## 0.0.2
 
  - **REFACTOR**: fix samples and examples.